requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohappyeyeballs==2.6.1
aiohttp~=3.9.0
aiohttp-socks<0.10
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
colorlog==6.9.0
cryptography==44.0.3
dotenv==0.9.9
frozenlist==1.5.0
g4f==0.5.2.5
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.4.3
nest-asyncio==1.6.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv==1.1.0
python-socks==2.7.1
pytz==2025.2
requests==2.32.3
rsa==4.9.1
Telethon==1.40.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
